pyproject.toml
[tool.poetry]
name = "ansara-task"
version = "0.1.0"
description = ""
authors = ["Demianight <komarovdemian@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
drf-nested-routers = "^0.94.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
Django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
